requirements.txt
Django
langchain
langchain-google-genai
python-dotenv
pypdf